Add spatial positioning and Zarr v2 compatibility for Neuroglancer
- Convert preprocess_1b and 1c to Zarr v2 format (numcodecs.Blosc)
- Add OME-NGFF translation transforms for spatial positioning
- Store basis vectors and bounding box in metadata for reconstruction
- Add visualize_simple.py with verbose mode and port checking
- Fix coordinate system issues in visualize_neuroglancer.py
- Update requirements and documentation

+++ src/preprocess_1_identify_bundles.py
@@ -42,7 +42,7 @@ def load_volume_downsampled(mat_file: Path, downsample: int = 4) -> Tuple[np.nda
     # MATLAB v7.3 files are HDF5-based
     with h5py.File(mat_file, 'r') as f:
         # Load final_lbl dataset
-        volume_full = f['final_lbl'][()].astype(np.uint32)
+        volume_full = f['final_lbl'][()]
 
         logger.info(f"Original shape: {volume_full.shape}")
